requirements.txt
Django==2.1.3
djangorestframework==3.9.0
drf-nested-routers==0.91
psycopg2-binary==2.7.6.1
pytz==2018.7
